refactor kind cli package layout
