+++ BUILD.bazel
@@ -1,0 +1,37 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["main.go"],
+    importpath = "k8s.io/test-infra/kind",
+    visibility = ["//visibility:private"],
+    deps = ["//kind/cmd/kind:go_default_library"],
+)
+
+go_binary(
+    name = "kind",
+    embed = [":go_default_library"],
+    visibility = ["//visibility:public"],
+)
+
+filegroup(
+    name = "package-srcs",
+    srcs = glob(["**"]),
+    tags = ["automanaged"],
+    visibility = ["//visibility:private"],
+)
+
+filegroup(
+    name = "all-srcs",
+    srcs = [
+        ":package-srcs",
+        "//kind/cmd/kind:all-srcs",
+        "//kind/hack:all-srcs",
+        "//kind/images:all-srcs",
+        "//kind/pkg/build:all-srcs",
+        "//kind/pkg/cluster:all-srcs",
+        "//kind/pkg/exec:all-srcs",
+    ],
+    tags = ["automanaged"],
+    visibility = ["//visibility:public"],
+)
