+++ cmd/kind/BUILD.bazel
@@ -1,17 +1,16 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library")
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "go_default_library",
-    srcs = ["main.go"],
+    srcs = ["kind.go"],
     importpath = "k8s.io/test-infra/kind/cmd/kind",
-    visibility = ["//visibility:private"],
-    deps = ["//kind/cmd/kind/cmd:go_default_library"],
-)
-
-go_binary(
-    name = "kind",
-    embed = [":go_default_library"],
     visibility = ["//visibility:public"],
+    deps = [
+        "//kind/cmd/kind/build:go_default_library",
+        "//kind/cmd/kind/create:go_default_library",
+        "//kind/cmd/kind/delete:go_default_library",
+        "//vendor/github.com/spf13/cobra:go_default_library",
+    ],
 )
 
 filegroup(
@@ -25,7 +24,9 @@ filegroup(
     name = "all-srcs",
     srcs = [
         ":package-srcs",
-        "//kind/cmd/kind/cmd:all-srcs",
+        "//kind/cmd/kind/build:all-srcs",
+        "//kind/cmd/kind/create:all-srcs",
+        "//kind/cmd/kind/delete:all-srcs",
     ],
     tags = ["automanaged"],
     visibility = ["//visibility:public"],
