+++ cmd/kind/delete/delete.go
